# docs/chatgpt-actions-openapi.yaml
openapi: 3.1.0
info:
  title: RFPmanager ChatGPT Actions
  version: 0.2.0
  description: API for listing, creating, reading, and updating RFPmanager opportunities.

servers:
  - url: https://rfpmanager.vercel.app

security:
  - chatgptActionsApiKey: []

paths:
  /api/chatgpt/rfps:
    get:
      operationId: listRfps
      summary: List all RFP opportunities
      description: >
        Use this operation for requests like "show all RFPs", "list RFPs",
        "pipeline", "opportunities", "active RFPs", "upcoming deadlines",
        or searches by client name or tender code. If the user names a client,
        tender code, or opportunity, use search to find the RFP instead of
        asking for a UUID.
      parameters:
        - name: search
          in: query
          required: false
          description: Client name, tender code, keyword, or opportunity detail to search for.
          schema:
            type: string
      responses:
        "200":
          description: List of RFP opportunities.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  rfps:
                    type: array
                    items:
                      $ref: "#/components/schemas/Rfp"

    post:
      operationId: createRfp
      summary: Create a new RFP opportunity
      description: Use this only when the user wants to add or create a new RFP record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfpInput"
      responses:
        "201":
          description: Created RFP.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  rfp:
                    $ref: "#/components/schemas/Rfp"

  /api/chatgpt/rfps/{id}:
    get:
      operationId: getRfp
      summary: Get one RFP opportunity by UUID
      description: >
        Use only when the RFP UUID is already known from an action response.
        If the user gives a client name, tender code, or description, call
        listRfps with search first. Do not ask for UUIDs unless search is
        ambiguous or empty.
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: RFP details.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  rfp:
                    $ref: "#/components/schemas/Rfp"

    patch:
      operationId: updateRfp
      summary: Update one RFP opportunity by UUID
      description: >
        Use only when updating an RFP whose UUID is known from an action
        response. If the user gives a client name, tender code, or description,
        call listRfps with search first. Do not ask for UUIDs unless search is
        ambiguous or empty.
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record to update.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfpUpdate"
      responses:
        "200":
          description: Updated RFP.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  rfp:
                    $ref: "#/components/schemas/Rfp"

  /api/chatgpt/rfps/{id}/documents:
    get:
      operationId: listRfpDocuments
      summary: List saved source document metadata for one RFP
      description: >
        Returns lightweight document metadata only. Does not return full
        Markdown, source file objects, download URLs, or large previews. To
        read content, call getRfpDocumentMarkdown with offset/limit.
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Lightweight saved document metadata.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  documents:
                    type: array
                    maxItems: 20
                    items:
                      $ref: "#/components/schemas/RfpDocumentSummary"
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
                  has_more:
                    type: boolean
                  next_offset:
                    type: ["integer", "null"]

  /api/chatgpt/rfps/{id}/source-files:
    get:
      operationId: listRfpSourceFiles
      summary: List original source files for one RFP
      description: >
        Use only when source file objects or temporary download URLs are needed.
        Do not call this before reading saved Markdown.
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Original source files with temporary download URLs when available.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  source_files:
                    type: array
                    items:
                      $ref: "#/components/schemas/RfpFileWithDownload"

  /api/chatgpt/rfps/{id}/documents/{documentId}:
    get:
      operationId: getRfpDocumentMarkdown
      summary: Get a bounded Markdown excerpt from one saved RFP document
      description: >
        Use this to read saved Markdown in chunks. Start with offset 0. If
        has_more is true, call again with next_offset. Keep limit modest unless
        the user explicitly asks for a larger extract.
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record, found through listRfps if needed.
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          description: The saved document UUID from listRfpDocuments.
          schema:
            type: string
            format: uuid
        - name: offset
          in: query
          required: false
          description: Character offset to start reading from.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum characters to return. The API caps this at 12000.
          schema:
            type: integer
            minimum: 1
            maximum: 12000
            default: 6000
      responses:
        "200":
          description: A bounded Markdown excerpt and pagination metadata.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  document:
                    $ref: "#/components/schemas/RfpDocumentExcerpt"

  /api/chatgpt/rfps/{id}/documents/convert:
    post:
      operationId: convertAndSaveRfpDocuments
      summary: Upload source files, convert them to Markdown, and save them to one RFP
      description: >
        Use this when a user provides tender source documents in ChatGPT and wants them saved in RFPmanager.
        The request must include openaiFileIdRefs; ChatGPT will populate each item with a temporary download_link.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAiFileUploadRequest"
      responses:
        "201":
          description: Saved source files and converted Markdown documents.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  saved:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      properties:
                        document:
                          $ref: "#/components/schemas/RfpDocumentSummary"
                        source_file:
                          $ref: "#/components/schemas/RfpFile"

  /api/chatgpt/rfps/{id}/summary:
    post:
      operationId: saveRfpSummary
      summary: Save a GPT-created summary for one RFP
      description: Use this after the GPT has reviewed saved Markdown and produced a summary. This does not call Groq.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - summary
              properties:
                summary:
                  type: string
                  description: The complete summary text to save against the RFP.
      responses:
        "200":
          description: Updated RFP with saved summary.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  rfp:
                    $ref: "#/components/schemas/Rfp"
                  summary:
                    type: ["string", "null"]

  /api/chatgpt/rfps/{id}/responses:
    get:
      operationId: listRfpResponses
      summary: List response files for one RFP
      description: Use this to check whether proposal response files have already been saved for an RFP.
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Saved response files.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  responses:
                    type: array
                    items:
                      $ref: "#/components/schemas/RfpFileWithDownload"

    post:
      operationId: saveRfpResponses
      summary: Upload GPT-created response files for one RFP
      description: >
        Use this after the GPT creates response files such as PDF, DOCX, or XLSX outputs.
        The request must include openaiFileIdRefs; ChatGPT will populate each item with a temporary download_link.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          description: The UUID of the RFP record.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAiFileUploadRequest"
      responses:
        "201":
          description: Saved response files.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  responses:
                    type: array
                    items:
                      $ref: "#/components/schemas/RfpFile"

components:
  securitySchemes:
    chatgptActionsApiKey:
      type: http
      scheme: bearer
      description: Shared API key sent as a bearer token in the Authorization header.

  schemas:
    Rfp:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        client_name:
          type: string
        status:
          type: string
          enum: ["Yes", "No", "TBD"]
        closing_date:
          type: ["string", "null"]
          description: ISO date in YYYY-MM-DD format, if available.
        tender_code:
          type: ["string", "null"]
        tender_link:
          type: ["string", "null"]
        gdrive_link:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contact_person:
          type: ["string", "null"]
        contact_phone:
          type: ["string", "null"]
        contact_email:
          type: ["string", "null"]
        document_links:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              name:
                type: string
              url:
                type: string
        summary_generated_at:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        pipeline_stage:
          type: string
          enum: ["Prospects", "Active", "Submitted", "Won", "Lost"]
        created_at:
          type: string

    OpenAiFileUploadRequest:
      type: object
      additionalProperties: false
      required:
        - openaiFileIdRefs
      properties:
        openaiFileIdRefs:
          type: array
          description: >
            One to ten files from the ChatGPT conversation. The parameter name must be openaiFileIdRefs.
            At runtime ChatGPT populates each item with id, name, mime_type, and a temporary download_link.
          minItems: 1
          maxItems: 10
          items:
            type: string

    RfpDocument:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        rfp_id:
          type: string
          format: uuid
        source_file_id:
          type: ["string", "null"]
          format: uuid
        title:
          type: string
        source_filename:
          type: ["string", "null"]
        source_type:
          type: string
          enum: ["docx", "pdf", "xlsx", "csv", "markdown"]
        markdown:
          type: string
        created_at:
          type: string

    RfpDocumentSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        rfp_id:
          type: string
          format: uuid
        source_file_id:
          type: ["string", "null"]
          format: uuid
        title:
          type: string
          maxLength: 200
        source_filename:
          type: ["string", "null"]
          maxLength: 255
        source_type:
          type: string
          enum: ["docx", "pdf", "xlsx", "csv", "markdown"]
        markdown_length:
          type: integer
        markdown_preview:
          type: string
          maxLength: 500
          description: First 500 characters only.
        created_at:
          type: string

    RfpDocumentExcerpt:
      allOf:
        - $ref: "#/components/schemas/RfpDocumentSummary"
        - type: object
          additionalProperties: false
          properties:
            markdown:
              type: string
            offset:
              type: integer
            limit:
              type: integer
            next_offset:
              type: ["integer", "null"]
            has_more:
              type: boolean

    RfpFile:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        rfp_id:
          type: string
          format: uuid
        kind:
          type: string
          enum: ["source", "response"]
        title:
          type: string
        original_filename:
          type: string
        mime_type:
          type: ["string", "null"]
        storage_path:
          type: string
        file_size_bytes:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_by:
          type: string
        created_at:
          type: string
        download_url:
          type: string
          format: uri

    RfpFileWithDownload:
      allOf:
        - $ref: "#/components/schemas/RfpFile"
        - type: object
          additionalProperties: false
          properties:
            download_url:
              type: ["string", "null"]
              format: uri
            download_error:
              type: string

    RfpInput:
      type: object
      additionalProperties: false
      required:
        - client_name
      properties:
        client_name:
          type: string
        status:
          type: string
          enum: ["Yes", "No", "TBD"]
          default: "TBD"
        closing_date:
          type: ["string", "null"]
          description: ISO date in YYYY-MM-DD format.
        closing_date_text:
          type: ["string", "null"]
          description: Human-readable closing date if an ISO date is not available.
        tender_code:
          type: ["string", "null"]
        tender_link:
          type: ["string", "null"]
        gdrive_link:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contact_person:
          type: ["string", "null"]
        contact_phone:
          type: ["string", "null"]
        contact_email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        pipeline_stage:
          type: string
          enum: ["Prospects", "Active", "Submitted", "Won", "Lost"]
          default: "Prospects"

    RfpUpdate:
      type: object
      additionalProperties: false
      properties:
        client_name:
          type: string
        status:
          type: string
          enum: ["Yes", "No", "TBD"]
        closing_date:
          type: ["string", "null"]
        tender_code:
          type: ["string", "null"]
        tender_link:
          type: ["string", "null"]
        gdrive_link:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contact_person:
          type: ["string", "null"]
        contact_phone:
          type: ["string", "null"]
        contact_email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        pipeline_stage:
          type: string
          enum: ["Prospects", "Active", "Submitted", "Won", "Lost"]